Restructured folders for charts
