+++ web/src/components/charts/ChartSummary/ChartSummary.tsx
