+++ web/src/components/charts/ChartSummary/index.tsx
