+++ web/src/components/charts/ZoomableTimeSeriesChart/index.tsx
@@ -16,5 +16,4 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-export * from './components';
-export { default as ChartSummary } from './ChartSummary';
+export { default } from './ZoomableTimeSeriesChart';
