+++ web/src/components/charts/constants.ts
